refactor(ts): migrar layout y Header completo a TypeScript
- AppLayout, Main, Aside, Footer
- Header, NavDesktop, NavMobile
- AdminMenu, GestionDropdown, MenuUsuarioDropdown, VendedoresDropdown
